requirements.txt
fastapi
uvicorn
python-docx
pdfminer.six
PyPDF2
scikit-learn